Revert "Bug Fixes"
This reverts commit 63c5c02bfc44e732399118d844906b0708e8e28c.

Above commit made a lot of unrelated changes without proper
explanations. One of the changes made for example causes inversal
of the behaviour to wash's -C flag (#106), and the other changes
seem highly controversial as well.
It is better to just revert them all and then with sufficient
time trying to reproduce and analyze what each of the changes tried
to achieve, and if indeed an improvement, reapply that specific change
including proper documentation (commit message) why that change was
made.