wash: ignore -C option, and default to using it
we now default to ignoring FCS error, and the option -C will just
be ignored. that way it stays compatible to scripts that use it,
while behaving as expected. the old behaviour of checking FCS can
be achieved by using the new -F flag.

the rationale for defaulting to ignoring FCS is that this is required
for compatibilty with ralink USB chips as pointed out by @kcdtv.