remove wash -C flag due to popular demand and default to ignore FCS errors
we now default to ignoring FCS errors, so there's no need anymore toa
set it explicitly.
the rationale for defaulting to ignoring FCS is that this is required
for compatibilty with ralink USB chips as pointed out by @kcdtv.
the maintainers did not deem it necessary to keep it as a no-op flag
for backwards compatibility, so let's just remove it.