remove usage of sqlite db
the following changes were merged into one

- remove get_db_ssid()

this seems to be one of the 2 only functions left that actually use
the sqlite db. it was used as a convenience function to look up
the ssid from the database when it was not found in a beacon.

- remove "auto-detect settings" feature

the purpose of this feature was to save the settings used for
attacking a specific AP, but i suppose it was almost universally
unused in the past because
1) most APs lock you out for an eternity after 3 wrong tries
2) those that don't you'll probably crack anyway with your first run.
3) most ppl are used to a workflow that add all the required command
   line arguments anyway
4) all automated tools add all the required command line arguments
   anyway.

so this is basically an almost-always unused lazy-ass switch that's
unheard of in most other programs, and as the only remaining feature
depending on sqlite, requiring sqlite installed just for this feature
is a heavy price to pay.

- remove unused function update_history

there's no functionality anymore that uses sql, so we dont need to
update the db anymore either.

- remove unused function get() from sql.c