import changes from t6x first commit vs vanilla reaver 1.4
whitespace changes have been omitted. unfortunately, the entire source
code of reaver is completely inconsistent in its usage of whitespace.
some parts are indented with tabs, others with spaces.
t6x' first commit was similarly inconsistent, and it changed *a lot*
of whitespace. would i have imported the whitespace changes as well,
it would be very hard to see what has changed, without usage of tools.

as the patch is very intrusive and added a number of bugs (for example
breaks option -g as found here:
https://github.com/t6x/reaver-wps-fork-t6x/issues/50#issuecomment-273802361 )
we add this commit only for documentary purposes and will revert it in
the follow-up commit, with the intention to study the changes and
forward-port only the necessary changes to implement the pixiedust
attack, and leave the bugs behind.