preliminary support for pixiedust attack
the values for the attack are taken from where the original t6x
change had them, but everything else is heavily refactored and
should be easy to trace back and improve to use different means
of executing the attack (i.e. use the pixiewps code directly rather
than starting it as an external command).