reaver: make -Z an alias for -K
the previous -K option automatically used the PIN found by pixie,
and tried to use it to retrieve the WPA key from the router.
there was a specific flag -Z to be used together with -K to suppress
this behaviour.
since this feature is currently not implemented, we just alias -Z to
-K for the time being.