Makefile.in: restore creation of reaver data dir
reaver traditionally stored it's (more or less useless) reaver.db
in /etc/reaver, plus *.wpc files for incomplete bruteforce sessions.
when we removed the reaver.db[0], nothing got installed into the data
dir so it was not created anymore. this had the effect that the
wpc files got always written into the current directory as a fallback.
people that want to force that files are always written into the
current directory, can use the configure option
--enable-savetocurrent[1] or just remove the /etc/reaver directory.

[0]: 46ccf2e268f74ca12c250157ae6017732e818f41
[1]: a3b840ec288b9024296ca57e8e9e492901a236bd