add missing usage text of pixiedust options
when the preliminary pixiedust attack was implemented, the usage text
was not added, so it might look to a user looking at the output as if
the attack was not implemented at all.
this was the case in issue #151.

until we make -K do the pixiedust attack plus the automatic wpa password
retrieval, keep a duplicated text entry for the -Z option, which is supposed
to run the pixiedust attack without password retrieval (so it is currently
doing the same).