wash: group getopt non-option arguments together
for compatibility with POSIX getopt, the recently introduced -j
option which doesn't take an argument, should be grouped together
with the other non-option arguments, since on the first occurence
of such an argument, it will consider all following arguments
non-option arguments too. this could be an issue with alternative
libc implementations.