wps_data_to_json: sanitize strings from probe response's WPS fields
a router of type TP-LINK TL-WA850RE sent as model number "2.0\x0d"
which had weird effects on the json string printed in the terminal.
any of those strings need to be sanitized, as a malicious AP could
include terminal escape sequences in its strings and so cause remote
code execution on the pentester's terminal.