wash: send more compliant probe requests
i encountered an AP of type

   "wps_manufacturer" : "TP-LINK",
   "wps_model_name"   : "TL-MR3220",
   "wps_model_number" : "1.0",
   "wps_device_name"  : "Wireless N 3G Router TL-MR3220"

which would not respond to probe requests sent by wash, but responded
immediately to "iw wlan0 scan". so i added some tags that most
standard clients usually send, and voila, it works now.

it stands to reason that wash should by default send only broadcast
probe requests, in order to keep a more stealthy profile.
as such, also the WPS tag should be emitted, since it is typically
not sent.