wash: if no channel number tag is found, assume current channel
some or all 5 GHz APs do not set tagged data for channel information.

since later on there's a check if channel == get_channel, this caused
wash to not print anything at all.

the code could be improved by getting the channel from the radiotap
header, but parsing that is rather complex, so we leave it for another
opportunity.