re-enable FCS validation
FCS was disabled in 31cd6da1c4ed55163443ea5da923b7fbb3e94f5d
due to wrong premise that it's not needed, when in fact it was just
too annoying having the terminal spammed with error messages using
a RALINK chipset.

now i turned off the terminal message, but turned the validation back
on. i implemented a new FCS check based on the field in the radiotap
header, as @alex-chan did when he submitted his OS X support patch,
however in a clean fashion, and used for any OS.

the reason this is needed on OSX is that they have very poor drivers
that sometimes submit only half packets, but with the BAD_FCS flag
set, which caused memory corruption and segfaults.

i assume the bad fcs check doesnt kick in now nearly as often as it
did when we disabled the check.