reaver: activate -n behaviour by default and obsolete the switch
most APs are conforming to the standard and send a NACK as expected
when the pin is wrong, but reaver defaulted to assuming the opposite.
this could cause reaver to wrongly assume a correct pin is incorrect
in non-optimal network conditions.

the old default behaviour to treat a timeout as NACK can be accessed
using the new -J / --timeout-is-nack switch.

closes #122