add compile-time switch to turn off htcaps for testing
since replaying 80211.n HT caps to the AP can affect several aspects
of its behaviour which have not yet been investigated in depth, it's
handy to have a compile time option to turn it off for testing.
if it turns out to produce further unexpected issues, we might want
to add a commandline option to reaver so it can be toggled on/off by
the user.
currently we do not have enough feedback to come to a definite
conclusion.