introduce resend timeout and resend packets when it is hit
the resend timeout defaults to 200 ms (1/5 sec). when it is hit, the
last packet sent will be sent again, until either the expected response
from the AP is received or the global timeout is hit.
this highly improves performance in high-congestion or long-distance
usage.

this currently only affects packets starting from EAPOL, but not those from the
association code, since they use their own resend logic.

example: assuming the timeout is set to 10 seconds, instead of sending M4 only
once, and then waiting for 10 seconds until the response of the AP is not received
and then failing with a timeout error, we now send the M4 every 1/5th sec during
that 10 second period.

credits: idea by @wiire-a