merge reaver and wash into same binary (busybox-style)
both programs use 95% of the same object code, i.e. they're mostly
equivalent. that means merging them into a single binary will save
almost 100% harddisk space.
reaver is now the main binary, while wash is a symlink to it.
main.c inspects argv[0] and then decides which program to run based
on the name the binary was started as.

closes #168