remove -D/--daemonize command line option
since SQLite support was removed, this option doesn't make any sense
anymore. with sqlite, a user could daemonize (i.e. send to background)
both reaver and wash, and the scan results from wash and an eventually
suceeding pin try from reaver would probably have made its way into the
database (but there didnt exist any way to see those results, without
resorting to the sqlite command line client...).

so without anything to back up the results automatically, losing the
text output of both reaver and wash defeats the entire purpose.
for people wanting to use the programs in the background, it is instead
recommended to use them with a terminal multiplexer such as GNU screen
or tmux, or via a script that redirects output to a file, and then
detaches from the controlling terminal...