pixiewps mode: implement automatic cracking and proper NACK usage
a couple routers started turning WPS off after a pixie attack.
research showed this happens only with reaver, not oneshot, so it's
either a defensive mechanism against reaver itself, or simply against
non-compliant WPS implementations.
as @feitoi pointed out, the best way to deal with that is to send
WSC NACK.

however, we can just as well continue the session if pixiewps succeeded
in a certain timeframe, and inject the found pin into the running
WPS transaction.

the new logic now works as follows:
pixiewps is started in a thread, which processes its output. if the
pin was cracked successfully during the normal transaction timeout
interval (-t parameter), reaver will try to inject the found pin.
if pixiewps fails, NACK is sent and the program terminated.
if we hit the timeout while pixiewps runs, we send NACK, but let
the pixiewps program running in the background until it terminates,
so the full output becomes visible, and finally terminate the program.

the thread is necessary because reading from the pixiewps process handle
is a blocking operation, and even if we'd detect a timeout, we still
couldn't get the main thread to interrupt its current read operation to
send the NACK immediately (and we can't do it from a SIGALRM signal
handler, because sending the NACK involves a lot of memory allocations
and other non-async-safe operations).

fixes #291