pixie: fix logic for wpc file update
reordering the code so we can safely exit after a pixie attack,
while having the wpc file updated. this prevents reaver from
retry-ing the pixiewps-found pin when the submission fails.
this allows the user to take a break and try it some minutes later
with -p manually, so he can circumvent time-based lock counters.