80211.c: simplify is_target()
is_target() had a confusing check for NULL_MAC: if the bssid
was set to it, it would return true.

the purpose of that is that wash by default isn't targetting a
specific bssid, so a target check needs to always succeed under
these circumstances. however it also allows to specify a specific
bssid, thus the is_target check.
the other site calling the function was from reaver code, which
always has the bssid set.
therefore i factored the null mac check out to the single site
it's needed, in wpsmon.c, which immediately makes the meaning
of both the check and the function much clearer.