add experimental new command line switch -u
as it is experimental, it is not yet being advertised in the
usage output.

the switch runs pixiewps with the -u uptime switch.
it is not actually a switch for pixiewps itself, but for
pixie-wrapper, which can use the timestamp for faster attacks
on hardcoded reset-dates in routers (in case the first regular
pixiewps pass fails).