Improve error reporting.
The return value of pcap_activate() is:

	0, if activation succeeded, with no warnings;
	a non-zero positive number, if activation succeded with a
	    warning - the number is a PCAP_WARNING_ indication of the
	    warning;
	a negative number, if activation failed - the number is a
	    PCAP_ERROR_ indication of the error.

We should not treat non-zero positive numbers as errors.  (Printing a
warning might be useful.)

We should treat non-zero negative numbers as errors, and translate it to
an error message using libpcap's own pcap_statustostr(), rather than
attempting to duplicate its mapping ourselves, as new error codes may
appear in the future, and using pcap_statustostr() future-proofs us
against that.  In addition, if the error is PCAP_ERROR, we should report
the return value of pcap_geterr(handle), as that will provide additional
information about that "generic" error.