wash: fix behaviour when AP cache is full
the old behaviour, which was changed recently, was to just not print
anything anymore when the internal cache of 512 APs was full.
b7c1d0e0ca0e0019b9fca8a2d28d1df64b7f07cb changed that to print everything
after the list was full.

a better way is to just clean the list when this happens and start
over, to prevent excessive spam in the output.